fix: app error when user is not signed in (#141)
there was an app error when a user trys to visit the dashboard but is not signed in; this isn't assumed to be nefarious behavior because there's a strong change a trusted user just has the /dashboard/* urls bookmarked or "memorized" in their browser search history. this code is going to push the user to the login screen if they aren't authenticated instead of throwing an error. I was getting too tr